refactor: Block, Node를 위한 폴더 추가 및 변경
