+++ src/main/java/org/bee/metro/core/block/application/BlockService.java
@@ -4,7 +4,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
-import org.bee.metro.core.block.domain.Block;
+import org.bee.metro.core.block.domain.block.Block;
 import org.springframework.stereotype.Service;
 
 @Service
