--- src/main/java/org/bee/metro/core/block/domain/BlockType.java
@@ -1,5 +1,0 @@
-package org.bee.metro.core.block.domain;
-
-public enum BlockType {
-    TEXT, IMAGE
-}
