+++ src/main/java/org/bee/metro/core/block/domain/block/BlockType.java
@@ -1,0 +1,5 @@
+package org.bee.metro.core.block.domain.block;
+
+public enum BlockType {
+    TEXT, IMAGE
+}
