+++ src/main/java/org/bee/metro/core/document/dto/DetailDocumentPayload.java
@@ -1,7 +1,7 @@
 package org.bee.metro.core.document.dto;
 
 import java.util.List;
-import org.bee.metro.core.block.domain.Block;
+import org.bee.metro.core.block.domain.block.Block;
 import org.bee.metro.core.document.domain.Document;
 import org.bee.metro.core.document.domain.Tag;
