+++ src/test/java/org/bee/metro/core/block/domain/BlockTest.java
@@ -5,6 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.util.UUID;
 import java.util.stream.Stream;
+import org.bee.metro.core.block.domain.block.Block;
+import org.bee.metro.core.block.domain.block.BlockType;
 import org.bee.metro.global.exception.type.BadRequestException;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
