+++ src/test/java/org/bee/metro/core/block/domain/NodeTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import java.util.Map;
 import java.util.UUID;
 import java.util.stream.Stream;
+import org.bee.metro.core.block.domain.node.Node;
 import org.bee.metro.global.exception.type.BadRequestException;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
